+++ packages/credential-provider-imds/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 4.0.7
+
+### Patch Changes
+
+- Updated dependencies [64e033f]
+  - @smithy/types@4.3.2
+  - @smithy/node-config-provider@4.1.4
+  - @smithy/property-provider@4.0.5
+  - @smithy/url-parser@4.0.5
+
 ## 4.0.6
 
 ### Patch Changes