+++ packages/eventstream-serde-universal/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.0.5
+
+### Patch Changes
+
+- Updated dependencies [64e033f]
+  - @smithy/types@4.3.2
+  - @smithy/eventstream-codec@4.0.5
+
 ## 4.0.4
 
 ### Patch Changes