+++ packages/util-retry/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.0.7
+
+### Patch Changes
+
+- Updated dependencies [64e033f]
+  - @smithy/types@4.3.2
+  - @smithy/service-error-classification@4.0.7
+
 ## 4.0.6
 
 ### Patch Changes