+++ packages/util-stream-node/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies [64e033f]
+  - @smithy/types@4.3.2
+  - @smithy/node-http-handler@4.1.1
+
 ## 4.0.7
 
 ### Patch Changes